requirements/base.txt
# Core dependencies for all environments
asgiref==3.11.1
bleach==6.3.0
Django==6.0.2
django-htmx==1.27.0
django-tagulous==2.1.1
Markdown==3.10.2
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
redis==7.2.1
sqlparse==0.5.5
tzdata==2025.3
tinycss2==1.5.1
webencodings==0.5.1

# Celery for async tasks (WEEK 5-6)
celery==5.6.2

# scikit-learn for TF-IDF vectorization (WEEK 5)
scikit-learn==1.8.0

# python-dotenv for environment variable management
python-dotenv==1.2.1
